Fix the CI lint issue
